Add detailed repo docs and download instructions
