+++ docs/4. repository/3. repo.sourcify.dev.mdx
@@ -1,0 +1,46 @@
+
+# repo.sourcify.dev 
+
+[repo.sourcify.dev](https://repo.sourcify.dev) is an interface to the Sourcify contract file repository `RepositoryV1`.
+
+The code is available at [sourcifyeth/h5ai-nginx](https://github.com/sourcifyeth/h5ai-nginx). For performance reasons, it is not possible to navigate the folders above the contract level. You need to know ahead the contract you are looking for.
+:::tip Lookup
+
+Instead of entering the chain, you can check an address over all chains at https://sourcify.dev/#/lookup
+:::
+
+The contracts are accessible under the following path format:
+
+```
+https://repo.sourcify.dev/contracts/:match/:chainId/:contractAddress
+```
+
+- `:match`: either `full_match` or `partial_match`
+- `:chainId`: EVM chain id `1` for Ethereum Mainnet, `5` Ethereum testnet Görli etc. See [chainlist.org](https://chainlist.org)
+- `:contractAddress`: e.g. `0x5ed4a410A612F2fe625a8F3cB4d70f197fF8C8be`
+
+### Examples
+
+Here are some example contracts:
+
+- https://repo.sourcify.dev/contracts/full_match/1/0x5ed4a410A612F2fe625a8F3cB4d70f197fF8C8be
+- https://repo.sourcify.dev/contracts/full_match/1/0xca2ad74003502af6B727e846Fab40D6cb8Da0035
+- https://repo.sourcify.dev/contracts/full_match/100/0x4f15a6e74CFC2F80D5967a8aB75F3c83D8043cF4
+- https://repo.sourcify.dev/contracts/partial_match/1/0xb857F1f4014A0C45C287667148417b6799Fe594E/
+- (staging) https://repo.staging.sourcify.dev/contracts/partial_match/69/0xb50cBeeFBCE78cDe83F184B275b5E80c4f01006A/sources/
+
+### View Source Code in Remix IDE
+
+It is possible to view the contract folder in the Remix IDE by clicking "View in Remix".
+
+Allow the Sourcify plugin on the next screen in Remix IDE (might take a while to load). The contract folder will be available under `verified-sources/<contract-address>` in the Remix file explorer.
+
+![Sourcify repository screenshot](/img/sourcify-repo.png)
+
+### Download folders
+
+You can download the whole folder by clicking on top left download icon.
+
+Alternatively you can select which files/folders to download by clicking the checkmarks, and click the download icon.
+
+![Sourcify repository screenshot](/img/sourcify-repo-download.png)
