+++ docs/4. repository/index.mdx
@@ -1,0 +1,16 @@
+# Contract Repository
+
+Sourcify stores the contracts in multiple storage backends and gives the option to choose which one to use. In short there are the following options:
+
+- `RepositoryV1`
+- `RepositoryV2`
+- `SourcifyDatabase`
+- `AllianceDatabase`
+
+For details see [Choosing the storage backend](https://github.com/ethereum/sourcify/tree/staging/services/server#choosing-the-storage-backend).
+
+## Download
+
+You can download the whole contract file repository in zips or the Sourcify database in Parquet format. Follow the guides in each page:
+- [Download RepositoryV2](/docs/repository/file-repositories/#download)
+- [Download SourcifyDatabase](/docs/repository/sourcify-database/#download)
