--- docs/99. Troubleshooting.md
@@ -1,26 +1,0 @@
----
-id: troubleshooting
-title: Troubleshooting
-slug: /troubleshooting
----
-
-# Troubleshooting
-
-## Error "o.filter is not a function" or "invalid chainIds"
-
-This happens when you add a new chain or move to a new branch with a new chain, but the build files don't contain the newly added chain. You should
-
-```
-npm install
-npx lerna run build
-```
-
-to create the build files with the added chains.
-
-## Reinstalling dependencies
-
-Switching branches, broken package links etc. may cause errors. In such cases you can try removing and installing the dependencies with:
-
-```
-npm run build:clean
-```
